Conflicts:
	application/controllers/admin/export.php
	application/controllers/admin/responses.php
	application/controllers/survey/index.php
	application/core/Survey_Common_Action.php
	application/helpers/admin/activate_helper.php
	application/helpers/admin/exportresults_helper.php
	application/helpers/admin/statistics_helper.php
	application/helpers/common_helper.php
	application/helpers/expressions/em_manager_helper.php
	application/helpers/frontend_helper.php
	application/helpers/qanda_helper.php
	application/helpers/replacements_helper.php
	application/helpers/update/updatedb_helper.php
	installer/sql/create-mssql.sql
	installer/sql/create-mysql.sql
	installer/sql/create-pgsql.sql

+++ application/controllers/survey/index.php
@@ -524,50 +524,52 @@ class index extends CAction {
         //Clear session and remove the incomplete response if requested.
         if (isset($move) && $move == "clearall")
         {
+            // delete the response but only if not already completed
             $s_lang = $_SESSION['survey_'.$surveyid]['s_lang'];
-            if (isset($_SESSION['survey_'.$surveyid]['srid']))
+            if (isset($_SESSION['survey_'.$surveyid]['srid']) && !Survey_dynamic::model($surveyid)->isCompleted($_SESSION['survey_'.$surveyid]['srid']))
             {
-                // find out if there are any fuqt questions - checked
-                $fieldmap = createFieldMap($surveyid,'short',false,false,$s_lang);
-                foreach ($fieldmap as $field)
-                {
-                    $q = $field['q'];
-                    if ($q->fileUpload() && !strpos($q->fieldname, "_filecount"))
+                // delete the response but only if not already completed
+                 $result= dbExecuteAssoc('DELETE FROM {{survey_'.$surveyid.'}} WHERE id='.$_SESSION['survey_'.$surveyid]['srid']." AND submitdate IS NULL");
+                if($result->count()>0){
+                    // find out if there are any fuqt questions - checked
+                    $fieldmap = createFieldMap($surveyid,'short',false,false,$s_lang); //AJS#
+                    foreach ($fieldmap as $field)
                     {
-                        $questions[] = $q;
-                    }
-                }
-
-                // if yes, extract the response json to those questions
-                if (isset($qid))
-                {
-                    $query = "SELECT * FROM {{survey_".$surveyid."}} WHERE id=".$_SESSION['survey_'.$surveyid]['srid'];
-                    $result = dbExecuteAssoc($query);
-                    foreach($result->readAll() as $row)
-                    {
-                        foreach ($questions as $q)
+                        $q = $field['q'];
+                        if ($q->fileUpload() && !strpos($q->fieldname, "_filecount"))
                         {
-                            $json = $row[$q->fieldname];
-                            if ($json == "" || $json == NULL)
-                                continue;
+                            if (!isset($questions)) { $questions = array(); }
+                            $qid[] = $q;
+                        }
+                    }
 
-                            // decode them
-                            $phparray = json_decode($json);
-
-                            foreach ($phparray as $metadata)
+                    // if yes, extract the response json to those questions
+                    if (isset($qid))
+                    {
+                        $query = "SELECT * FROM {{survey_".$surveyid."}} WHERE id=".$_SESSION['survey_'.$surveyid]['srid'];
+                        $result = dbExecuteAssoc($query);
+                        foreach($result->readAll() as $row)
+                        {
+                            foreach ($qid as $question)
                             {
-                                $target = Yii::app()->getConfig("uploaddir")."/surveys/".$surveyid."/files/";
-                                // delete those files
-                                unlink($target.$metadata->filename);
+                                $json = $row[$q->fieldname];
+                                if ($json == "" || $json == NULL)
+                                    continue;
+
+                                // decode them
+                                $phparray = json_decode($json);
+
+                                foreach ($phparray as $metadata)
+                                {
+                                    $target = Yii::app()->getConfig("uploaddir")."/surveys/".$surveyid."/files/";
+                                    // delete those files
+                                    unlink($target.$metadata->filename);
+                                }
                             }
                         }
                     }
+                    // done deleting uploaded files
                 }
-                // done deleting uploaded files
-
-
-                // delete the response but only if not already completed
-                dbExecuteAssoc('DELETE FROM {{survey_'.$surveyid.'}} WHERE id='.$_SESSION['survey_'.$surveyid]['srid']." AND submitdate IS NULL");
 
                 // also delete a record from saved_control when there is one
                 dbExecuteAssoc('DELETE FROM {{saved_control}} WHERE srid='.$_SESSION['survey_'.$surveyid]['srid'].' AND sid='.$surveyid);
